% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_box}
\alias{bar_box}
\title{Distribution of count values}
\usage{
bar_box(
  x,
  interplot = FALSE,
  seg = NULL,
  lab = "Number of months food insecure"
)
}
\arguments{
\item{x}{vector with the variable of interest}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{seg}{segmentation of households (if any)}

\item{lab}{the label on the axis}
}
\description{
This function show the distribution of count values
such as the HDDS score or the number of months with food insecurity
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

#distribution of number of month with food insecurity
bar_box(hhinfo$foodshortage_count)

#distribution of HDDS score per large region
bar_box(hhinfo$hdds_score, lab="HDDS", seg=hhinfo$large_region)

}
\keyword{boxplot}
