% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_div}
\alias{bar_div}
\title{Barplot of the crop or livestock adoption}
\usage{
bar_div(
  tab,
  idlist = NULL,
  th = 10,
  seg = NULL,
  interplot = FALSE,
  colcat = "#1F77B4",
  marx = 4
)
}
\arguments{
\item{tab}{either crop or lvst table with columns hhid and name}

\item{idlist}{full list of households id}

\item{th}{threshold to show all items with more than th (default=10)}

\item{seg}{segmentation of households (if any)}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{colcat}{colors of the items}

\item{marx}{extra margin for plotting the crop names (when not interactive plot)}
}
\description{
This function make a barplot showing the most commun
crop or livestock in the dataset
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

bar_div(crop, idlist=hhinfo$hhid)
bar_div(lstk, idlist=hhinfo$hhid)

}
\keyword{adoption}
\keyword{barplot,}
