% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_gender}
\alias{bar_gender}
\title{Gender control over farm production}
\usage{
bar_gender(tab, interplot = FALSE, top = 10)
}
\arguments{
\item{tab}{the dataframe with information on farm products}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{top}{select how many categories will be shown (by default 10)}
}
\description{
This function show the gender control over consumed and sold farm products
}
\examples{
data(hhdb_rhomis)

#plot gender control on crop
bar_gender(hhdb_rhomis$crop)

#plot gender control on livestock product
bar_gender(hhdb_rhomis$lstk_prod)

}
\keyword{gender}
