% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_hist}
\alias{bar_hist}
\title{Distribution of numerical values as histogram}
\usage{
bar_hist(
  x,
  breaks = c(0, 0.5, 1, 2, 5, 10),
  lab = NULL,
  xunit = "ha",
  pal = NULL,
  interplot = FALSE,
  seg = NULL
)
}
\arguments{
\item{x}{vector with the variable of interest}

\item{breaks}{breaks to build the histogram, by default 6 breaks for positive values}

\item{lab}{the name of the classes (should be one length shorter than break)}

\item{xunit}{the label of the axis}

\item{pal}{colors of the numerical categories  (should be one length shorter than break)}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{seg}{segmentation of households (if any)}
}
\description{
This function show the distribution of quantities
such as land cultivated, livestock herd size, family size, ...
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

#distribution of land cultivated
bar_hist(hhinfo$land_cultivated_ha)

#land cultivated per HDDS score
bar_hist(hhinfo$land_cultivated_ha, seg=hhinfo$hdds_score)

}
\keyword{hist}
