% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_income}
\alias{bar_income}
\title{Show the months with food insecurity}
\usage{
bar_income(
  tab,
  qmax = 0.95,
  marx = 6,
  seg = NULL,
  interplot = FALSE,
  maxbar = 100
)
}
\arguments{
\item{tab}{the dataframe with information on the income activities, usually hhinfo}

\item{qmax}{percentile that is not shown on y axis}

\item{marx}{extra margin for plotting the names (when not interactive plot)}

\item{seg}{segmentation of households (if any)}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{maxbar}{number of bars to be plotted. If too high, the barplot is hardly readable.}
}
\description{
This function show the months with food insecurity
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

#plot the source of income
bar_income(hhinfo)

#plot the source of income per large region
bar_income(hhinfo, seg=hhinfo$large_region)

}
\keyword{income}
\keyword{sources}
