% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_months}
\alias{bar_months}
\title{Show the months with food insecurity}
\usage{
bar_months(x, sep = " ", interplot = FALSE, lab = "Month with food insecurity")
}
\arguments{
\item{x}{the vector containing the months with food shortage}

\item{sep}{the separation between months, by default " "}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{lab}{the label on the axis}
}
\description{
This function show the months with food insecurity
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

#plot the month with food shortage score
bar_months(hhinfo$foodshortage_months)

}
\keyword{food}
\keyword{months}
\keyword{security}
