% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_score}
\alias{bar_score}
\title{Distribution of scores}
\usage{
bar_score(
  tab,
  score = "HDDS",
  pal = NULL,
  marx = 6,
  rm0 = FALSE,
  interplot = FALSE,
  seg = NULL
)
}
\arguments{
\item{tab}{the dataframe with information on the score}

\item{score}{name of the score and prefix of columns of interest}

\item{pal}{colors of the categories}

\item{marx}{extra margin for plotting the names (when not interactive plot)}

\item{rm0}{remove households with a score of 0 (are they hidden NA?)}

\item{interplot}{decide if the plot is interactive (with plotly) or not}

\item{seg}{segmentation of households (if any)}
}
\description{
This function show the distribution of scores that are built on multiple categories
such as the HDDS or the FIES
}
\examples{
data(hhdb_rhomis)
attach(hhdb_rhomis)

#plot HDDS score
bar_score(hhinfo, score="HDDS")

#plot FIES per large region
bar_score(hhinfo, score="FIES", seg=hhinfo$large_region)

}
\keyword{diet}
\keyword{diversity}
