% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{bestname}
\alias{bestname}
\title{Find the best match from a list of known names}
\usage{
bestname(x, y, warn = FALSE)
}
\arguments{
\item{x}{is the original names}

\item{y}{is the list of known names}

\item{warn}{whether to show a warning if missing names}
}
\value{
the names that best match the known names
}
\description{
Find the best match from a list of known names
}
\examples{
bestname(c("bean", "sweetpotatoes", "apples"),c("apple", "beans", "sweet_potatoes"))

}
