% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_farm_prod.R
\name{calc_farm_prod}
\alias{calc_farm_prod}
\title{Calculate farm household characteristics}
\usage{
calc_farm_prod(crop, lstk, lstk_prod, hhinfo, conv_tlu, conv_energy)
}
\arguments{
\item{crop}{crop information}

\item{lstk}{livestock information}

\item{lstk_prod}{livestock production}

\item{hhinfo}{information on households}

\item{conv_tlu}{conversion from livestock categories to tlu}

\item{conv_energy}{energy conversion of product from kg to kcal}
}
\description{
This function summarize farm household activities
based on information about crop, livestock and household
}
\keyword{summary}
