% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cleanname}
\alias{cleanname}
\title{Clean names}
\usage{
cleanname(x)
}
\arguments{
\item{x}{is a character vector}
}
\value{
the cleaned names without brackets, and spaces
}
\description{
Clean names
}
\examples{
cleanname(c("maize", "sweet potatoes", "apple(tree)"))

}
