% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Count_NA0}
\alias{Count_NA0}
\title{Count number of 0, negative values or NA}
\usage{
Count_NA0(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
a number of elements with NA, 0 or negative values
}
\description{
Count number of 0, negative values or NA
}
\examples{
Count_NA0(c(-99, NA, -1:9))

}
