% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Count_Spaces}
\alias{Count_Spaces}
\title{Count the number of unique elements in a string}
\usage{
Count_Spaces(x, sep = " ", na.rm = TRUE)
}
\arguments{
\item{x}{is a string}

\item{sep}{separator between two elements of the string}

\item{na.rm}{define if NA are removed from the count}
}
\value{
a number of unique element
}
\description{
This function evaluate the number of unique elements in a string
}
\examples{
Count_Spaces("maize NA banana maize beans")

}
