% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhdb_rhomis}
\alias{hhdb_rhomis}
\title{RHoMIS dataset}
\format{
\describe{
  \item{hhdb_rhomis}{a farmhousehold object}
}
}
\source{
\url{https://www.rhomis.org}
}
\usage{
hhdb_rhomis
}
\description{
A sample dataset containing household information from RHoMIS dataset
}
\references{
van Wijk, M., Hammond, J., Gorman, L., Adams, S., Ayantunde, A., Baines, D., ... & Yameogo, V. (2020). The Rural Household Multiple Indicator Survey, data from 13,310 farm households in 21 countries. Scientific Data, 7(1), 46.
}
\keyword{datasets}
