% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lunique}
\alias{lunique}
\title{Number of unique elements of a vector}
\usage{
lunique(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{define if NA are removed from the count}
}
\value{
number of unique elements
}
\description{
This function returns the number of unique elements
in a vector
}
\examples{
lunique(c("a", 5, NA, 5, month.abb[1]))

}
