% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{nan0}
\alias{nan0}
\title{Transform NaN to 0 (for bar_gender plot)}
\usage{
nan0(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
a vector with NaN transformed into 0
}
\description{
Transform NaN to 0 (for bar_gender plot)
}
