% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{nan100}
\alias{nan100}
\title{Transform NaN to 100 (for bar_gender plot)}
\usage{
nan100(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
a vector with NaN transformed into 100
}
\description{
Transform NaN to 100 (for bar_gender plot)
}
