% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{NAto0}
\alias{NAto0}
\title{Transform NA and negative values to 0}
\usage{
NAto0(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
a vector with NA and negative values transformed into 0
}
\description{
Transform NA and negative values to 0
}
\examples{
NAto0(c(-99, NA, -1:9))

}
