% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{NAto1}
\alias{NAto1}
\title{Transform NA and negative values to 1 (for conversion factors)}
\usage{
NAto1(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
a vector with NA and negative values transformed into 1
}
\description{
Transform NA and negative values to 1 (for conversion factors)
}
\examples{
NAto1(c(-99, NA, -1:9))

}
