% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{new_farmhousehold}
\alias{new_farmhousehold}
\title{Create a new farmhousehold object}
\usage{
new_farmhousehold(crop, lstk, lstk_prod, hhinfo, conv_tlu, conv_energy)
}
\arguments{
\item{crop}{crop table as data.frame}

\item{lstk}{livestock table as data.frame}

\item{lstk_prod}{livestock production table as data.frame}

\item{hhinfo}{household information table as data.frame}

\item{conv_tlu}{numeric vector with tlu conversion factors}

\item{conv_energy}{numeric vector with energy conversion factors}
}
\description{
Create a new farmhousehold object
}
