% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{panel.cor.m}
\alias{panel.cor.m}
\title{Correlation plot for pairs}
\usage{
panel.cor.m(x, y, digits = 2, method = "pearson")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{digits}{precision of the correlation}

\item{method}{correlation method ("pearson" (default), "kendall", or "spearman")}
}
\value{
plot with the correlation coefficient
}
\description{
Correlation plot for pairs
}
