% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pie_seg}
\alias{pie_seg}
\title{Show the pie chart of the number of households per group}
\usage{
pie_seg(x, interplot = FALSE)
}
\arguments{
\item{x}{the vector containing the groups}

\item{interplot}{decide if the plot is interactive (with plotly) or not}
}
\description{
This function show the threshold of a numeric variable
}
\examples{
data(hhdb_rhomis)

#plot the source of income
pie_seg(hhdb_rhomis$hhinfo$large_region)


}
\keyword{pie}
