% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{Select the threshold for numeric variables}
\usage{
plot_density(x, th, lab = "", corskew = TRUE, pal = c("red", "blue"))
}
\arguments{
\item{x}{the vector containing the variable of interest}

\item{th}{the threshold}

\item{lab}{the label on the axis}

\item{corskew}{correct the highly skewed variable}

\item{pal}{color of the below and above threshold}
}
\description{
This function show the threshold of a numeric variable
}
\examples{
data(hhdb_rhomis)

#plot the source of income
plot_density(hhdb_rhomis$hhinfo$hh_size_mae, th=5)

}
\keyword{density}
\keyword{threshold}
