% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{print.farmhousehold}
\alias{print.farmhousehold}
\title{Print short summary of the farmhousehold data}
\usage{
\method{print}{farmhousehold}(x)
}
\arguments{
\item{x}{farmhousehold object}
}
\description{
Print short summary of the farmhousehold data
}
