% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{punique}
\alias{punique}
\title{Collapse unique elements of a vector}
\usage{
punique(x, sep = " ", na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{sep}{separator between two elements}

\item{na.rm}{define if NA are removed from the count}
}
\value{
one string with all unique elements together
}
\description{
This function returns a string with
the unique elements collapsed together in a single string
}
\examples{
punique(c("maize", NA, "banana", "maize", "beans"))

}
