% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{saveCSV_farmhousehold}
\alias{saveCSV_farmhousehold}
\title{Export the farmhousehold information as four csv files}
\usage{
saveCSV_farmhousehold(x, name, dirfile = NULL)
}
\arguments{
\item{x}{farmhousehold object}

\item{name}{name of the dataset}

\item{dirfile}{directory of the output file}
}
\description{
Export the farmhousehold information as four csv files
}
\examples{
data(hhdb_rhomis)
# saveCSV_farmhousehold(hhdb_rhomis, "rhomis")
}
