% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{saveRDS_farmhousehold}
\alias{saveRDS_farmhousehold}
\title{Export the farmhousehold information as rds files}
\usage{
saveRDS_farmhousehold(x, name, path = NULL)
}
\arguments{
\item{x}{farmhousehold object}

\item{name}{name of the dataset}

\item{path}{directory of the output file}
}
\description{
Export the farmhousehold information as rds files
}
\examples{
data(hhdb_rhomis)
saveRDS_farmhousehold(hhdb_rhomis, "rhomis")
}
