% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{scale01}
\alias{scale01}
\title{Scale values between 0 and 1}
\usage{
scale01(x, ...)
}
\arguments{
\item{x}{is a numeric vector}

\item{...}{additional elements that are passed to min() and max()}
}
\value{
the scaled vector with value between 0 and 1
}
\description{
Scale values between 0 and 1
}
\examples{
scale01(c(-99, NA, -1:9))

}
