% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{simplifynames}
\alias{simplifynames}
\title{Simplify a list of crop names with known names}
\usage{
simplifynames(x, conv_name, warn = TRUE)
}
\arguments{
\item{x}{is the original names}

\item{conv_name}{is the list of known names}

\item{warn}{add warning if the name is missing from conv_name}
}
\value{
the vector wuth the unique crop names
}
\description{
Simplify a list of crop names with known names
}
\examples{
simplifynames(c("bean", "sweetpotatoes", "apples"),c("apple", "beans", "sweet_potatoes"))

}
