% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sum_NAto0}
\alias{sum_NAto0}
\title{Sum all values apart of negative values and NA}
\usage{
sum_NAto0(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
the sum of the values of x
}
\description{
Sum all values apart of negative values and NA
}
\examples{
sum_NAto0(c(-99, NA, -1:9))==sum(1:9)

}
