% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{tmap_ind}
\alias{tmap_ind}
\title{Create a tmap of households gps location}
\usage{
tmap_ind(x, var, j = 0.05, logS = TRUE, interplot = TRUE)
}
\arguments{
\item{x}{farmhousehold object or dataframe with, at least, gps_lat, gps_lon and the variable of interest}

\item{var}{the name of the variable to be shown}

\item{j}{parameter to jitter the coordinates}

\item{logS}{whether the highly skewed variable are log(x+1) transformed}

\item{interplot}{decide if the plot is interactive or not}
}
\description{
This function creates a tmap and highlight one selected variable
}
\examples{
data(hhdb_rhomis)

#plot the source of income
tmap_ind(hhdb_rhomis, "large_region")

}
