% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{uniqueNA}
\alias{uniqueNA}
\title{Unique values that are not NA}
\usage{
uniqueNA(x)
}
\arguments{
\item{x}{is a vector}
}
\value{
the unique elements that are not NA
}
\description{
Unique values that are not NA
}
\examples{
uniqueNA(c("a", 5, NA, 5, month.abb[1]))

}
