% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{update.farmhousehold}
\alias{update.farmhousehold}
\title{Update the crop and livestock production statistics per household}
\usage{
\method{update}{farmhousehold}(x)
}
\arguments{
\item{x}{farmhousehold object}
}
\description{
Update the crop and livestock production statistics per household
}
\examples{
data(hhdb_rhomis)
# increase all price by 50\%
hhdb_rhomis$crop$income_lcu <- hhdb_rhomis$crop$income_lcu*1.5
#update the summary
hhdb_rhomis_highprice <- update.farmhousehold(hhdb_rhomis)
median(hhdb_rhomis_highprice$hhinfo$crop_value_lcu/hhdb_rhomis$hhinfo$crop_value_lcu, na.rm=TRUE)
}
