% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{validate_farmhousehold}
\alias{validate_farmhousehold}
\title{Validate a farmhousehold object}
\usage{
validate_farmhousehold(x)
}
\arguments{
\item{x}{farmhousehold S3 object}
}
\description{
Validate a farmhousehold object
}
