% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhdb_rhomis}
\alias{hhdb_rhomis}
\title{RHoMIS dataset}
\format{
\describe{
  \item{hhinfo}{Household information}
  \item{crop}{Crop production}
  \item{lstk}{Livestock herd}
  \item{lstk_prod}{Production from livestock}
}
}
\source{
\url{https://www.rhomis.org}
}
\usage{
hhdb_rhomis
}
\description{
A dataset containing household information from RHoMIS dataset
}
\references{
Van Wijk et al.
}
\keyword{datasets}
