% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhousehold.R
\name{saveCSV_hhinfo}
\alias{saveCSV_hhinfo}
\title{Export the household table information as csv}
\usage{
saveCSV_hhinfo(x, name, dirfile = NULL)
}
\arguments{
\item{x}{farmhousehold object}

\item{name}{name of the dataset}

\item{dirfile}{directory of the output file}
}
\description{
Export the household table information as csv
}
\examples{
data(hhdb_rhomis)
# saveCSV_hhinfo(hhdb_rhomis, "rhomis_hhinfo")
}
